changinging int in drug

+++ models/drug.py
@@ -14,7 +14,7 @@ class Drug(BaseModel, Base):
     __tablename__ = "drugs"
     id = Column(Integer, primary_key=True)
     stock_id = Column(Integer, ForeignKey("stocks.id"), nullable=False)
-    consultation_id = Column(Integer(6), ForeignKey("consultations.id"), nullable=False)
+    consultation_id = Column(Integer, ForeignKey("consultations.id"), nullable=False)
     name = Column(String(60), nullable=False)
     drug_state = Column(String, Enum('gutt', 'tab', 'oint', 'caps'))
     quantity = Column(Integer)
